refactor: modularize protocol and logging
