+++ jarvis/protocols/executor.py
@@ -8,8 +8,8 @@ import time
 
 from ..logger import JarvisLogger
 from ..agents.agent_network import AgentNetwork
-from . import Protocol
-from .mongo_logger import ProtocolUsageLogger, generate_protocol_log
+from .models import Protocol
+from ..loggers.protocol_logger import ProtocolUsageLogger, generate_protocol_log
 
 
 class ProtocolExecutor:
